Fix invalid cast for binds to levels; Fix BindManager.Update not updating properties -BindManager.cs  Properties try to type switch on helper classes when the underlying class is not derived from them -CachedEnvelope.cs:  Removed the helper classes as they no longer served a purpose;  Used reflection to invoke another method which has a gremlin inside if you try to invoke it directly (wtf?).  AddAmount returns current value back to caller after mutating the envelope -Channel.cs:  Set initial TL to first value of scaled envelope if a bind exists for it so there's no longer a "pop" sound between initial TL and the first tick of the bind -Test2.cs:  Fix integer-based TrackerEnvelopes return values making an invalid cast to float when attempting to Xerp() the values

+++ ui/main/voice/Test2.cs
@@ -216,7 +216,7 @@ public class Test2 : Label
                     float val = log_scale? e switch
                     {  //Levels are exponential in nature, so remap them using exponential interpolation
                         TrackerEnvelope<float> f => (float)Tools.Xerp((float)f.MinValue, (float)f.MaxValue, 1.0-pt.y),
-                        TrackerEnvelope<int> i => (float)Tools.Xerp((float)i.MinValue, (float)i.MaxValue, 1.0-pt.y),
+                        TrackerEnvelope<int> i => (float)Tools.Xerp((int)i.MinValue, (int)i.MaxValue, 1.0-pt.y),
                         _ => default
                     }: e switch  {  //Remap the input value from 0-1 to our binds bounds using linear interpolation instead
                         TrackerEnvelope<float> f => (float)Tools.Remap(pt.y, 0, 1, (float)f.MinValue, (float)f.MaxValue),
@@ -233,7 +233,7 @@ public class Test2 : Label
                     val = log_scale? e switch
                     {  //Levels are exponential in nature, so remap them using exponential interpolation
                         TrackerEnvelope<float> f => (float)Tools.Xerp((float)f.MinValue, (float)f.MaxValue, 1.0-pt.y),
-                        TrackerEnvelope<int> i => (float)Tools.Xerp((float)i.MinValue, (float)i.MaxValue, 1.0-pt.y),
+                        TrackerEnvelope<int> i => (float)Tools.Xerp((int)i.MinValue, (int)i.MaxValue, 1.0-pt.y),
                         _ => default
                     }: e switch  {  //Remap the input value from 0-1 to our binds bounds using linear interpolation instead
                         TrackerEnvelope<float> f => (float)Tools.Remap(pt.y, 0, 1, (float)f.MinValue, (float)f.MaxValue),
@@ -305,7 +305,7 @@ public class Test2 : Label
             float val = log_scale? e switch
             {  //Levels are exponential in nature, so remap them using exponential interpolation
                 TrackerEnvelope<float> f => (float)Tools.Xerp((float)f.MinValue, (float)f.MaxValue, 1.0-pt.y),
-                TrackerEnvelope<int> i => (float)Tools.Xerp((float)i.MinValue, (float)i.MaxValue, 1.0-pt.y),
+                TrackerEnvelope<int> i => (float)Tools.Xerp((int)i.MinValue, (int)i.MaxValue, 1.0-pt.y),
                 _ => default
             }: e switch  {  //Remap the input value from 0-1 to our binds bounds using linear interpolation instead
                 TrackerEnvelope<float> f => (float)Tools.Remap(pt.y, 0, 1, (float)f.MinValue, (float)f.MaxValue),
